# Conflicts:
#	src/main/java/sjht/erp/home/Mapper/HomeMapper.java
#	src/main/java/sjht/erp/home/controller/HomeApiController.java
#	src/main/java/sjht/erp/home/service/HomeService.java
#	src/main/java/sjht/erp/home/service/HomeServiceImpl.java
#	src/main/resources/mappers/home/homeMapper.xml
#	src/main/resources/static/css/menu/grid.css
#	src/main/resources/static/css/menu/home.css
#	src/main/resources/templates/menu/home.html

+++ src/main/java/sjht/erp/interceptor/Interceptor.java
@@ -21,7 +21,7 @@ public class Interceptor implements HandlerInterceptor {
 
         if(request.getAttribute("auth")!=null){
             Authentication authentication = (Authentication) request.getAttribute("auth");
-            Long userId = ((EmployeeDto) authentication.getPrincipal()).getEmpno();
+            int userId = ((EmployeeDto) authentication.getPrincipal()).getEmpno();
 
             EmployeeDto empNo = loginService.findByUserId(userId);
